+++ db/migrate/20200805015444_create_meal_foods.rb
@@ -7,25 +7,25 @@ class CreateMealFoods < ActiveRecord::Migration[6.0]
 
       # figure out which of these need to be included here and which I can pull from mealFoods.attributes.food on the frontend
       t.integer :calories
-      t.integer :total_fat
-      t.integer :saturated_fat
-      t.integer :polyunsaturated_fat
-      t.integer :monounsaturated_fat
-      t.integer :trans_fat
-      t.integer :cholesterol
-      t.integer :sodium
-      t.integer :total_carbohydrate
-      t.integer :dietary_fiber
-      t.integer :total_sugars
-      t.integer :added_sugars
-      t.integer :sugar_alcohols
-      t.integer :protein
-      t.integer :vitamin_a
-      t.integer :vitamin_c
-      t.integer :vitamin_d
-      t.integer :calcium
-      t.integer :iron
-      t.integer :potassium
+      t.float :total_fat
+      t.float :saturated_fat
+      t.float :polyunsaturated_fat
+      t.float :monounsaturated_fat
+      t.float :trans_fat
+      t.float :cholesterol
+      t.float :sodium
+      t.float :total_carbohydrate
+      t.float :dietary_fiber
+      t.float :total_sugars
+      t.float :added_sugars
+      t.float :sugar_alcohols
+      t.float :protein
+      t.float :vitamin_a
+      t.float :vitamin_c
+      t.float :vitamin_d
+      t.float :calcium
+      t.float :iron
+      t.float :potassium
 
       t.timestamps
     end
